services/vanna/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
vanna==0.3.4
groq==0.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.4.2
pandas==2.1.4
numpy<2.0
